pczt: Changes to be usable by Keystone firmware
